feat: separate domains in folders
