+++ src/commands/recipient/create-recipient-user.ts
@@ -3,9 +3,9 @@ import { Register } from '@boostercloud/framework-types'
 import { RecipientCreated as RecipientCreated } from '../../events/recipient/recipient-created'
 import { AuthService } from '../../services/auth-service'
 import { TypeOfIdentityDocument } from '../../common/type-of-identity-document'
-import { RecipientUserNotFoundInFirebaseError as RecipientNotFoundInFirebaseError } from '../../common/recipient-user-not-registered-in-firebase-error'
 import { UserRole } from '../../common/user-role'
 import { getUserId } from '../../common/user-utils'
+import { RecipientUserNotFoundInFirebaseError } from '../../common/errors/recipient-user-not-registered-in-firebase-error'
 
 @Command({
   authorize: 'all',
@@ -25,7 +25,7 @@ export class CreateRecipient {
     const userId: string = getUserId(register)
     await AuthService.setRole(userId, UserRole.Recipient).catch((error) => {
       console.log(error)
-      throw new RecipientNotFoundInFirebaseError(userId)
+      throw new RecipientUserNotFoundInFirebaseError(userId)
     })
 
     register.events(
